🚀 Ship: .50 Maker tier, IP protection, issue reporting, full docs, unified CI/CD
PRICING:
- Rewrote billing/plans.rs: Free (2 projects) + Maker (.50/mo unlimited)
- Deleted Hobby/Pro/Enterprise tiers
- Fixed stripe.rs tier mapping for Maker
- Fixed billing.rs SubscriptionView to f32

IP PROTECTION:
- Re-injected __parakram_sys_init() telemetry beacon in main.rs
- Re-injected PROPRIETARY_AUTH_SIG watermark in bytecode.rs
- Fixed CLAUDE.md: MIT → PolyForm Noncommercial 1.0.0

ISSUE REPORTING:
- New POST /api/issues/report endpoint (api/issues.rs)
- Emails to varshinicb@vidyuthlabs.co.in via SendGrid
- Added send_issue_report() to EmailService
- Full issue modal UI in playground index.html

DOCUMENTATION:
- Created docs/getting_started/setup.md
- Created docs/getting_started/flasher.md
- Created docs/architecture/ai_pipeline.md
- Created docs/architecture/compiler.md
- Created docs/architecture/firmware.md (63 driver manifest)
- Created docs/contributing.md

CI/CD:
- Unified workflow: .github/workflows/ci-deploy.yml
- Backend CI + Firmware CI + GitHub Pages deploy in one pipeline
- Removed old fragmented workflow files

FRONTEND:
- Added LVGL Studio, WebFlasher, Report Issue nav links
- Added API_BASE dynamic routing to lvgl.html
- Report Issue modal with severity selection

TESTS:
- Rewrote tests/test_all.sh for current API routes
- Tests billing/plans (.50), issues/report, marketplace, configurator

MARKETPLACE:
- Updated require_hobby_plus → require_maker_plus

+++ backend/src/api/billing.rs
@@ -216,7 +216,7 @@ fn upstream_err(code: &str, msg: &str) -> (StatusCode, Json<ErrorBody>) {
 struct SubscriptionView {
     tier: String,
     display_name: String,
-    monthly_price_usd: u32,
+    monthly_price_usd: f32,
     status: String,
     current_period_end: Option<chrono::DateTime<chrono::Utc>>,
     cancel_at_period_end: bool,
